Tweaks and new layer2
Layer 2, which is the symbols layer, got shuffled around and
reorganized. It also gained (), {}, and [], which were previously bound
to left / right shift, and left / right alt, respectively.

Shift and alt keys are now only shift and alt keys, while tab and enter
retain their lctrl / rctrl status too.

Changed left space -> right space from backspace to enter, because
backspace is already easy to access. So is enter, but it's still one
fewer reason to use the pinky.

If / when I move to homerow mods, and if I manage to move tab to some
other good place, I can probably stop using the pinkies completely
outside of the base layer of alphas and symbols